PHP Started - Search Bar Implemented
To get a start with the database functions and PHP, I added a search bar near the top of the main page (placement not final obviously).

If the user doesn't type anything in the search bar or makes it blank, all reviews are shown. Otherwise, sql queries will look through the movieReview and showReview tables for media with titles that match the query.

The output is really rough for now; just echos with the info found in the database. But the queries work.

Another quick note: The login button now links to the login page, and I've uploaded this version of our project to Hostinger so I could test the database functions.