Php API Interactions Added
I used the TMDB API with php to display posters when searching for movies or tv shows.

After our discussion in class, I've realized that we may have to use JS over PHP for some of the db interactions. I am now going to see what I can do about having reviews and the associated posters displayed with js instead. Unfortunate to potentially have to redo some of this, but I understand the fundamentals now.

Also did a little bit of cleanup and neatening of the process.php file's layout.